Added script that allows footsteps sounds
It is a bit jank it works better with one button then with several current solution is to hide it under the in sounds of the factory. I  looking online I think using animation events for this if a walk cycle is added would make for the best sound quality and feel, but would force us to make a walk cycle which may not be practical. Code likely needs more tuning but I want to get started on the background noise for the therapy room first.